putting things

+++ kagura/myjoblib.py
@@ -3,6 +3,7 @@ joblib wrapper: put files in a directory
 """
 
 from sklearn.externals import joblib
+import os
 
 def load(filename):
     filepath = os.path.join(filename, 'index')
@@ -13,6 +14,10 @@ def load(filename):
 
 def dump(value, filename):
     import os
+    if os.path.exist(filename):
+        # user specified an existing filename, avoid overwriting
+        dump(value, filename + "_NEW")
+        return
     os.makedirs(filename)
     filepath = os.path.join(filename, 'index')
     joblib.dump(value, filepath)
